drawing outlines of 1 flower using Beziers
Using bezier curves to draw the outlines of a flower that will be evenly distributed radially around a circle.
Bezier seemed to be the most intuitive to me as I am used to illustrators pen tool. It is just a slow process to draw complex symmetrical shapes and I spent a bunch of time going back and forth between my hand drawn image and p5 to make sure I was correctly redrawing it in p5.

To make it easier to see what is going on I put all the bezier curves into its own function and then called that function.

The next step is to create the fill colours for this flower. Then I can finally put it into the radial distribution.