modifying flower code so that it can be filled
When drawing the flower I made the mistake of using bezier instead of bezierVertex inside of beginShape. This meant that I couldn't fill the shapes properly as they weren't connected to each other. luckily Qodo could convert what I had already drawn to beginShapes, so I didn't have to start over from the beginning but I did then have to add in manually the connecting lines that create a fillable shape. It took a bit of time but once this was done it was easy to fill the shapes.

next step is to test the radial distribution.