getting flowers to face outwards from radial distribution centre
making the flowers face outwards from centre of circle. This one was a little bit annoying to troubleshoot. It kept doing weird things because I had drawn the flower at 100,100 instead of 0,0. Once i figured out that I had to rotate around 0,0  instead of 100,100 it worked.

the next step is to add if statement to it, to either change the colour of every even flower or change the direction.