Adding if statement to flip every even flower
testing out if statement for flipping every even flower 180 degrees. Then setting up another if statement using true/false to toggle it on and off. This will give me some added flexibility later on when generating the 9 images.

the next step is too also create an if statement for colours. - making flexible colours, potentially by using an array to call from.

Then will move on to the other flowers or fillagry in the composition.