Setting up colour arrays to create adjustable colours in the flowers
setting up colour arrays for both the light colours and dark colours. testing with just blue and green in this situation but will add more options.

arrays on p5.js website was helpful for this.

then setting up a variable that selects between the array. eg. lightClrSelect- essentially can just toggle between, 0 and 1 to change between blue and green versions.

This allows me to change the colours easily with just one variable change and I can also change the light colours and dark colours independently so can get lots of flexibility of colour palettes.

next step is to add a bunch of colours I like and then can move on to other shapes.