Starting to draw the center flower
Starting to draw the center flower, working on the center dots and a toggable switch between a center diamond and center ellipse at very center of canvas using if statements.

for the center dots I used a nested for loop, reusing and modifying code from my radially distributed flowers to create the dots following a circular path. I then used a nested y loop to create the concentric rings. I learnt how to create concentric circles inside of each other through Patt Vira's multiple circles with for loop video . I changed it so that instead of circles in a single loop like in the video I am using a nested for loop so I can use my radially distributed ellipses instead. I had to add in some extra variables to make it work but it was much easier than manually drawing every dot.

circle for loop video:
https://www.youtube.com/watch?v=5z_qEe7QxjU

I also added in a colour array for the background so that I can easily switch between bg colours.

the next steps are to finish the center flower, figure out how to have the center flower change size, amount etc using if statement or arrays, and set up a bunch of colours I like in the global colour arrays.