drawing the petal for the center flower
drawing a petal for the center flower, using beginShape, bezierVertex etc. Will then radially distribute it using the same code as flower 2 so I can have a variable for the amount of petals in the center flower.

Also added in a bg colour for the center flowers center dots, using a filled ellipse.