starting to radially distribute the petals- happy accidents
A little happy accident occured when starting to radially distribute the petals. Because the petals weren't at 0,0 the previous code from flower 2, didn't rotate them outwards and instead the petals follow the tangent of the circle. I thought this was a cool geometric effect that could be used for an if statement between a geometric shape in the middle and a flower in the middle to create flexibility.

Will adjust it and then set up and if statement to toggle between the two options.