adjusting Radial distribution of petals to make it work for flower
Tweaking the for loop rotation , radius , amount of petals etc settings to get the petals for the flower looking good

Next step is too add in some additional petals or triangles to match the complexity /aesthetic of the other flowers in the sketch

Also Setting up if statement to toggle between the geometric petal pattern  of the petals and flower petals. Not sure if I will do it this way, I might clean it up by setting up seperate functions for the center flower and geometric center and then toggling between the two.