Adding in corner shapes for tiling
Adding in the corner shapes for tiling. For now just using the same motif as middle to test. Middle of motif is set to 0,0. 0,200. 200,0. and 200,200. That way they line up to create a full motif when tiled.

next step is too come up with the different variations for the middle motif and the corner motifs. Then create if statements or arrays to select between them.