drawing filigree to add to composition
drawing filigree using beginShape, bezierVertex etc.

Also testing out how to mirror in p5. I learnt how to use the scale function to invert either the x or y xis from the coding trains scale video:

https://www.youtube.com/watch?v=pkHZTWOoTLM&t=379s

Next step is to radially distribute or mirror it to create patterns with the filigree.

Once that is done I need to setup a a way of turning all motifs on and off independently.

Also cleaned up a bunch of redundant code and put things into their own functions to make it easier for me going forward since there is a lot of code in the sketch now.