Setting up if statements to select between 4 center motif options
Setting up if statements to select between 5 center motif options. Using if statement plus a variable ( centerObject )which I can select between the 5 options by changing from 0, to 1 to 2 to 3 etc.

Also adjusted the vertically flipped filigree to look different than horizontally one to give me more flexibility.

the next step is to sort out the same selection setup for the corners and add in filigree options for the corners.

Then it is just adding the colour options.