cleaning up push pop error and cleaning up
Spent a couple hours trying to fix a weird translate error that would cause the centerFiligree1 function to mess up the corner filigrees in the last code.

Finally figured out that it was because I was missing 1 push function  in the that function and it was not treating everything as independent systems.

 Also cleaned up the code some more and added in more comments to explain what everything does since the code is so big now.

I adjusted the final filigree element to give me more variance and then I changed the names of the functions to be easier to read.

FInal step is to add the extra colour palettes and then clean up one more time and bounce out images.